FIX: Sprites
Kain, might be able to help with the whys.
I always need the sprite data to be done as an interleaved bitmap for a sprite to display.

You also need seperate data for each sprite, since the sprite writes information to the first part of the sprite data, which is why you data is 2 extra lines in size.

On OCS/ECS the palette for sprites are colours are in the last 16 colours, (even on a 16 colour screen).

                16  Unused   00  |
                 17  Color 1  01  |
                 18  Color 2  10  |-- Sprites 0 and 1
                 19  Color 3  11 _|
                 20  Unused   00  |
                 21  Color 1  01  |
                 22  Color 2  10  |-- Sprites 2 and 3
                 23  Color 3  11 _|
                 24  Unused   00  |
                 25  Color 1  01  |
                 26  Color 2  10  |-- Sprites 4 and 5
                 27  Color 3  11 _|
                 28  Unused   00  |
                 29  Color 1  01  |
                 30  Color 2  10  |-- Sprites 6 and 7
                 31  Color 3  11 _|